Feature/elasticsearch #32

+++ src/main/java/com/swyp/catsgotogedog/content/repository/ContentRepository.java
@@ -1,8 +1,7 @@
 package com.swyp.catsgotogedog.content.repository;
 
+import com.swyp.catsgotogedog.content.domain.entity.Content;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.swyp.catsgotogedog.content.domain.entity.Content;
-
-public interface ContentRepository extends JpaRepository<Content, Long> {
+public interface ContentRepository extends JpaRepository<Content, Integer> {
 }
